refactor(seeds): relocate cluster masters into world.seeds.game_content (roadmap 3.2, #1220)
Move the seed cluster master modules (battles, challenges, characters, checks,
clash, combat, conditions, items, magic, social) from
integration_tests/game_content/ to world/seeds/game_content/, repointing
clusters.py and intra-package imports to the new home. integration_tests/
game_content/ keeps a thin per-module compatibility facade (explicit re-exports)
so every existing test import keeps working unchanged.
